// frontend/src/live/generated/live.ts
export interface SubscribeRequest {
    /**
     * @generated from protobuf field: string session_id = 1
     */
    sessionId: string;
}
/**
 * @generated from protobuf message sigmap.live.v1.LiveEvent
 */
export interface LiveEvent {
    /**
     * @generated from protobuf field: string session_id = 1
     */
    sessionId: string;
    /**
     * @generated from protobuf oneof: payload
     */
    payload: {
        oneofKind: "detections";
        /**
         * @generated from protobuf field: sigmap.live.v1.DetectionBatch detections = 2
         */
        detections: DetectionBatch;
    } | {
        oneofKind: "device";
        /**
         * @generated from protobuf field: sigmap.live.v1.Heartbeat device = 3
         */
        device: Heartbeat;
    } | {
        oneofKind: "config";
        /**
         * @generated from protobuf field: sigmap.live.v1.ConfigPush config = 4
         */
        config: ConfigPush;
    } | {
        oneofKind: "fleet";
        /**
         * @generated from protobuf field: sigmap.live.v1.FleetSnapshot fleet = 5
         */
        fleet: FleetSnapshot;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message sigmap.live.v1.DetectionBatch
 */
export interface DetectionBatch {
    /**
     * @generated from protobuf field: string batch_id = 1
     */
    batchId: string;
    /**
     * @generated from protobuf field: string device_id = 2
     */
    deviceId: string;
    /**
     * @generated from protobuf field: repeated sigmap.live.v1.LocatedDetection detections = 3
     */
    detections: LocatedDetection[];
}
/**
 * @generated from protobuf message sigmap.live.v1.LocatedDetection
 */
export interface LocatedDetection {
    /**
     * @generated from protobuf field: string mac = 1
     */
    mac: string;
    /**
     * @generated from protobuf field: string device_type = 2
     */
    deviceType: string;
    /**
     * @generated from protobuf field: optional string ssid = 3
     */
    ssid?: string;
    /**
     * @generated from protobuf field: optional string bt_name = 4
     */
    btName?: string;
    /**
     * @generated from protobuf field: int32 signal_dbm = 5
     */
    signalDbm: number;
    /**
     * @generated from protobuf field: int32 channel = 6
     */
    channel: number;
    /**
     * @generated from protobuf field: string encryption = 7
     */
    encryption: string;
    /**
     * @generated from protobuf field: string detected_at = 8
     */
    detectedAt: string;
    /**
     * @generated from protobuf field: string location_flag = 9
     */
    locationFlag: string;
    /**
     * @generated from protobuf field: optional double lat = 10
     */
    lat?: number;
    /**
     * @generated from protobuf field: optional double lon = 11
     */
    lon?: number;
    /**
     * @generated from protobuf field: optional string vendor_name = 12
     */
    vendorName?: string;
}
/**
 * @generated from protobuf message sigmap.live.v1.Heartbeat
 */
export interface Heartbeat {
    /**
     * @generated from protobuf field: string device_id = 1
     */
    deviceId: string;
    /**
     * @generated from protobuf field: string at = 2
     */
    at: string;
    /**
     * @generated from protobuf field: string status = 3
     */
    status: string;
    /**
     * @generated from protobuf field: optional int32 current_channel = 4
     */
    currentChannel?: number;
    /**
     * @generated from protobuf field: optional int32 battery_pct = 5
     */
    batteryPct?: number;
    /**
     * @generated from protobuf field: bool gps_fix = 6
     */
    gpsFix: boolean;
    /**
     * @generated from protobuf field: int32 detections_buffered = 7
     */
    detectionsBuffered: number;
    /**
     * @generated from protobuf field: int32 detections_sent_total = 8
     */
    detectionsSentTotal: number;
}
/**
 * @generated from protobuf message sigmap.live.v1.ConfigPush
 */
export interface ConfigPush {
    /**
     * @generated from protobuf field: string device_id = 1
     */
    deviceId: string;
    /**
     * @generated from protobuf field: string push_id = 2
     */
    pushId: string;
    /**
     * @generated from protobuf field: string status = 3
     */
    status: string;
}
/**
 * @generated from protobuf message sigmap.live.v1.FleetSnapshot
 */
export interface FleetSnapshot {
    /**
     * @generated from protobuf field: repeated sigmap.live.v1.Heartbeat devices = 1
     */
    devices: Heartbeat[];
}
// @generated message type with reflection information, may provide speed optimized methods
class SubscribeRequest$Type extends MessageType<SubscribeRequest> {
    constructor() {
        super("sigmap.live.v1.SubscribeRequest", [
            { no: 1, name: "session_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<SubscribeRequest>): SubscribeRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sessionId = "";
        if (value !== undefined)
            reflectionMergePartial<SubscribeRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SubscribeRequest): SubscribeRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string session_id */ 1:
                    message.sessionId = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SubscribeRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string session_id = 1; */
        if (message.sessionId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.sessionId);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message sigmap.live.v1.SubscribeRequest
 */
export const SubscribeRequest = new SubscribeRequest$Type();
// ...
export const LiveEvent = new LiveEvent$Type();
// ...
export const DetectionBatch = new DetectionBatch$Type();
// ...
export const LocatedDetection = new LocatedDetection$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Heartbeat$Type extends MessageType<Heartbeat> {
    constructor() {
        super("sigmap.live.v1.Heartbeat", [
            { no: 1, name: "device_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "at", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "status", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "current_channel", kind: "scalar", opt: true, T: 5 /*ScalarType.INT32*/ },
            { no: 5, name: "battery_pct", kind: "scalar", opt: true, T: 5 /*ScalarType.INT32*/ },
            { no: 6, name: "gps_fix", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 7, name: "detections_buffered", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 8, name: "detections_sent_total", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<Heartbeat>): Heartbeat {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.deviceId = "";
        message.at = "";
        message.status = "";
        message.gpsFix = false;
        message.detectionsBuffered = 0;
        message.detectionsSentTotal = 0;
        if (value !== undefined)
            reflectionMergePartial<Heartbeat>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Heartbeat): Heartbeat {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string device_id */ 1:
                    message.deviceId = reader.string();
                    break;
                case /* string at */ 2:
                    message.at = reader.string();
                    break;
                case /* string status */ 3:
                    message.status = reader.string();
                    break;
                case /* optional int32 current_channel */ 4:
                    message.currentChannel = reader.int32();
                    break;
                case /* optional int32 battery_pct */ 5:
                    message.batteryPct = reader.int32();
                    break;
                case /* bool gps_fix */ 6:
                    message.gpsFix = reader.bool();
                    break;
                case /* int32 detections_buffered */ 7:
                    message.detectionsBuffered = reader.int32();
                    break;
                case /* int32 detections_sent_total */ 8:
                    message.detectionsSentTotal = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Heartbeat, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string device_id = 1; */
        if (message.deviceId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.deviceId);
        /* string at = 2; */
        if (message.at !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.at);
        /* string status = 3; */
        if (message.status !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.status);
        /* optional int32 current_channel = 4; */
        if (message.currentChannel !== undefined)
            writer.tag(4, WireType.Varint).int32(message.currentChannel);
        /* optional int32 battery_pct = 5; */
        if (message.batteryPct !== undefined)
            writer.tag(5, WireType.Varint).int32(message.batteryPct);
        /* bool gps_fix = 6; */
        if (message.gpsFix !== false)
            writer.tag(6, WireType.Varint).bool(message.gpsFix);
        /* int32 detections_buffered = 7; */
        if (message.detectionsBuffered !== 0)
            writer.tag(7, WireType.Varint).int32(message.detectionsBuffered);
        /* int32 detections_sent_total = 8; */
        if (message.detectionsSentTotal !== 0)
            writer.tag(8, WireType.Varint).int32(message.detectionsSentTotal);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message sigmap.live.v1.Heartbeat
 */
export const Heartbeat = new Heartbeat$Type();
// ...
export const ConfigPush = new ConfigPush$Type();
// ...
export const FleetSnapshot = new FleetSnapshot$Type();
/**
 * @generated ServiceType for protobuf service sigmap.live.v1.LiveStream
 */
export const LiveStream = new ServiceType("sigmap.live.v1.LiveStream", [
    { name: "Subscribe", serverStreaming: true, options: {}, I: SubscribeRequest, O: LiveEvent }
]);